requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=0.32.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beartype=0.18.5=pypi_0
beautifulsoup4=4.12.3=pypi_0
braceexpand=0.1.7=pypi_0
bs4=0.0.2=pypi_0
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.17.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=4.4.2=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
dukpy=0.4.0=pypi_0
einops=0.6.0=pypi_0
einx=0.3.0=pypi_0
ema-pytorch=0.5.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.15.4=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.1=pypi_0
frozendict=2.4.4=pypi_0
fsspec=2024.6.1=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.66.0=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.24.0=pypi_0
idna=3.7=pypi_0
imageio=2.34.2=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.0=pyh41d4057_0
jax=0.4.13=pypi_0
jaxlib=0.4.13=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
keras=2.13.1=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.6=pypi_0
lpip=1.0.7=pypi_0
lpips=0.1.4=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mergedeep=1.3.4=pypi_0
ml-dtypes=0.2.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
mutf8=1.0.6=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.3=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=1.4.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pdf2image=1.17.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.0=py38h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
prompt_toolkit=3.0.47=hd8ed1ab_0
protobuf=4.25.4=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pypac=0.16.4=pypi_0
pyparsing=3.1.2=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
requests-file=2.1.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.8.1=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.10.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.1=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.13.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-gan=2.1.0=pypi_0
tensorflow-hub=0.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
tensorflow-probability=0.21.0=pypi_0
termcolor=2.4.0=pypi_0
tf-keras=2.15.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.14=h39e8969_0
tldextract=5.1.2=pypi_0
tokenizers=0.19.1=pypi_0
torch=2.3.1=pypi_0
torchmetrics=1.4.1=pypi_0
torchtyping=0.1.4=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.42.4=pypi_0
triton=2.3.1=pypi_0
typeguard=4.3.0=pypi_0
typing-extensions=4.5.0=pypi_0
urllib3=2.2.2=pypi_0
vector-quantize-pytorch=1.14.8=pypi_0
wandb=0.17.5=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webdataset=0.2.86=pypi_0
werkzeug=3.0.4=pypi_0
wfdb=4.1.2=pypi_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
